Cicd batch (#47)
* cicd batch and pytest

* fix tasts

* fix test v2

* fix test3

* fix testing - just csv for now

* fix testing - warp db engine

* fix test batch

* reorder transform and predict

* fix batch cicd

+++ src/models/batch/beam_processing.py
@@ -19,38 +19,41 @@ import sys
 project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..'))
 sys.path.append(project_root)
 
-from config.config import DRIVER_CLASS_NAME, PASSWORD, JDBC_URL, USERNAME, INPUT_TABLE, OUTPUT_TABLE, INPUT_TYPE, OUTPUT_TYPE, JDBC_URL
+from config.config import result_columns, DRIVER_CLASS_NAME, PASSWORD, JDBC_URL, USERNAME, INPUT_TABLE, OUTPUT_TABLE, INPUT_TYPE, OUTPUT_TYPE, JDBC_URL
+IS_TESTING = os.environ.get('TESTING',"False") == 'True'
 
 current_dir = os.path.dirname(os.path.abspath(__file__))
 
 INPUT_DIR = os.path.join(current_dir, '..', '..', '..', 'data', 'batch_input')
 MODEL_PATH = os.path.join(current_dir, '..', '..', '..', 'models', 'churn_model.pickle')
 OUTPUT_DIR = os.path.join(current_dir, '..', '..', '..', 'data', 'batch_results')
-
+if IS_TESTING:
+    INPUT_DIR = 'csv'
+    OUTPUT_DIR = 'csv'
 whylabs_org_id = os.environ.get("WHYLABS_ORG_ID")
 whylabs_api_key = os.environ.get("WHYLABS_API_KEY")
 whylabs_dataset_id = os.environ.get("WHYLABS_DATASET_ID")
+if not IS_TESTING:
+    engine = create_engine(JDBC_URL)
+    Session = sessionmaker(bind=engine)
 
-engine = create_engine(JDBC_URL)
-Session = sessionmaker(bind=engine)
+    def discard_incomplete(data):
+        return len(data['Contract']) > 0 and len(data['tenure']) > 0
 
-def discard_incomplete(data):
-    return len(data['Contract']) > 0 and len(data['tenure']) > 0
+    metadata = MetaData()
+    output_table = Table(OUTPUT_TABLE, metadata,
+        Column('customerID', String, primary_key=True, nullable=False),
+        Column('TotalCharges', Float),
+        Column('MonthlyCharges', Float),
+        Column('tenure', Integer),
+        Column('Contract', String),
+        Column('PhoneService', Integer),
+        Column('prediction', Float),
+        Column('timestamp', DateTime, default=datetime.utcnow)
+    )
 
-metadata = MetaData()
-output_table = Table(OUTPUT_TABLE, metadata,
-    Column('customerID', String, primary_key=True, nullable=False),
-    Column('TotalCharges', Float),
-    Column('MonthlyCharges', Float),
-    Column('tenure', Integer),
-    Column('Contract', String),
-    Column('PhoneService', Integer),
-    Column('prediction', Float),
-    Column('timestamp', DateTime, default=datetime.utcnow)
-)
-
-metadata.drop_all(engine, tables=[output_table])
-metadata.create_all(engine)
+    metadata.drop_all(engine, tables=[output_table])
+    metadata.create_all(engine)
 
 class DiscardIncompleteDoFn(beam.DoFn):
     def process(self, element):
@@ -90,11 +93,17 @@ class TransformData(beam.DoFn):
         else:
             dataset['PhoneService'] = 0
         
-        if 'Contract' in dataset.columns:
-            dataset['Contract'] = dataset['Contract'].fillna('Unknown')
-        else:
-            dataset['Contract'] = 'Unknown'
-        
+        # if 'Contract' in dataset.columns:
+        #     dataset['Contract'] = dataset['Contract'].fillna('Unknown')
+        # else:
+        #     dataset['Contract'] = 'Unknown'
+        contract_dummies = pd.get_dummies(dataset['Contract'])
+        dataset = pd.concat([dataset, contract_dummies], axis=1)
+
+        for contract_type in ['Month-to-month', 'One year', 'Two year']:
+            if contract_type not in dataset.columns:
+                dataset[contract_type] = 0
+
         return [dataset.to_dict('records')[0]]
 
 def get_csv_headers(file_path):
@@ -112,15 +121,8 @@ class Predict(beam.DoFn):
             return []
         
         dataset = pd.DataFrame([element])
-        contract_dummies = pd.get_dummies(dataset['Contract'])
-        dataset = pd.concat([dataset, contract_dummies], axis=1)
 
-        
-        for contract_type in ['Month-to-month', 'One year', 'Two year']:
-            if contract_type not in dataset.columns:
-                dataset[contract_type] = 0
-        
-        result_columns = ['TotalCharges', 'Month-to-month', 'One year', 'Two year', 'PhoneService', 'tenure']
+        # result_columns = ['TotalCharges', 'Month-to-month', 'One year', 'Two year', 'PhoneService', 'tenure']
         prediction = self.model.predict(dataset[result_columns])
         element['prediction'] = float(prediction[0])
         return [element]
@@ -215,8 +217,9 @@ def run(input_type=INPUT_TYPE, output_type=OUTPUT_TYPE, db_url=None, input_table
     if "csv" in input_type or "both" in input_type:
         for file in os.listdir(INPUT_DIR):
             dataset = pd.read_csv(os.path.join(INPUT_DIR, file))
-            results = why.log(dataset)
-            results.writer("whylabs").write()
+            if not IS_TESTING:
+                results = why.log(dataset)
+                results.writer("whylabs").write()
             
             p = beam.Pipeline(options=options)
             if file.endswith(".csv"):
@@ -241,11 +244,12 @@ def run(input_type=INPUT_TYPE, output_type=OUTPUT_TYPE, db_url=None, input_table
                 
                 
                 #return collected_data | beam.Map(apply_whylogs)
-                try:
-                    results = collected_data | beam.Map(apply_whylogs)
-                    results.writer("whylabs").write()
-                except:
-                    print("eror in writing the results to whylabs")
+                if not IS_TESTING:
+                    try:
+                        results = collected_data | beam.Map(apply_whylogs)
+                        results.writer("whylabs").write()
+                    except:
+                        print("eror in writing the results to whylabs")
                 os.remove(DATA_PATH)
                 print(f"File {DATA_PATH} deleted.")
     
@@ -261,12 +265,13 @@ def run(input_type=INPUT_TYPE, output_type=OUTPUT_TYPE, db_url=None, input_table
         )
         
         write_results(output_type, data)
-        try:
-            # Log the data to WhyLabs
-            results = why.log(data[['customerID','TotalCharges', 'Month-to-month', 'One year', 'Two year', 'PhoneService', 'tenure', 'prediction', 'timestamp']])
-            results.writer("whylabs").write()
-        except:
-            print("eror in writing the results to whylabs")
+        if not IS_TESTING:
+            try:
+                # Log the data to WhyLabs
+                results = why.log(data[['customerID','TotalCharges', 'Month-to-month', 'One year', 'Two year', 'PhoneService', 'tenure', 'prediction', 'timestamp']])
+                results.writer("whylabs").write()
+            except:
+                print("eror in writing the results to whylabs")
         # Run the pipeline
 
         p.run().wait_until_finish()
